% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdjustProb.R
\name{AdjustProb}
\alias{AdjustProb}
\title{AdjustProb}
\usage{
AdjustProb(v, MaxElem)
}
\arguments{
\item{v}{a numeric vector}

\item{MaxElem}{number of elements}
}
\description{
AdjustProb
}
\keyword{internal}
