% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CICA_data.R
\docType{data}
\name{CICA_data}
\alias{CICA_data}
\title{Example data for CICA}
\format{
An object of class \code{list} of length 3.
}
\usage{
data(CICA_data)
}
\description{
Example data for CICA
}
\examples{
data(CICA_data)
}
\keyword{datasets}
