% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusf.R
\name{clusf}
\alias{clusf}
\title{Random clustering generation}
\usage{
clusf(nBlocks, nClus)
}
\arguments{
\item{nBlocks}{integer: number of objects}

\item{nClus}{integer: number of clusters}
}
\value{
a numeric vector indicating the clustering
}
\description{
Internal function for CICA package
}
\keyword{internal}
