% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleData.R
\docType{data}
\name{ExampleData}
\alias{ExampleData}
\title{Example data for CICA}
\format{
An object of class \code{list} of length 9.
}
\usage{
data(ExampleData)
}
\description{
Example data for CICA
}
\examples{
data(ExampleData)
}
\keyword{datasets}
