% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindRationalStarts.R
\name{FindRationalStarts}
\alias{FindRationalStarts}
\title{Compute rational starts for CICA}
\usage{
FindRationalStarts(
  DataList,
  nComp,
  nClus,
  scale = TRUE,
  center = TRUE,
  verbose = TRUE,
  pseudo = NULL
)
}
\arguments{
\item{DataList}{a list of matrices}

\item{nComp}{number of ICA components to extract}

\item{nClus}{number of clusters}

\item{scale}{scale each matrix to have an equal sum of squares}

\item{center}{mean center matrices}

\item{verbose}{print output to console}

\item{pseudo}{default is \code{NULL}}
}
\value{
dataframe with rational and pseudo-rational starts
}
\description{
Compute rational starts for CICA
}
\references{
Durieux, J., & Wilderjans, T. F. (2019). Partitioning subjects based on high-dimensional fMRI data: comparison of several clustering methods and studying the influence of ICA data reduction in big data. Behaviormetrika, 46(2), 271-311.
}
