% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateRandomClustering.R
\name{GenerateRandomClustering}
\alias{GenerateRandomClustering}
\title{Generate random clustering}
\usage{
GenerateRandomClustering(nElement, nClust, Prob = NULL)
}
\arguments{
\item{nElement}{a positive integer number of elements to be clusterd}

\item{nClust}{a positive integer: Number of clusters}

\item{Prob}{numeric vector: Portion of elements in each cluster}
}
\value{
a numeric vector indicating the clustering
}
\description{
Internal function for CICA package
}
\keyword{internal}
