% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenRanStarts.R
\name{GenRanStarts}
\alias{GenRanStarts}
\title{Generate random starts}
\usage{
GenRanStarts(
  RanStarts,
  nClus,
  nBlocks,
  ARIlim = 0.2,
  itmax = 1000,
  verbose = FALSE
)
}
\arguments{
\item{RanStarts}{number of randomstarts to generate}

\item{nClus}{number of clusters}

\item{nBlocks}{number of objects}

\item{ARIlim}{maximal value of adjusted Rand Index}

\item{itmax}{maximum number of iterations used to find suitable random starts}

\item{verbose}{boolean that indicates whether the output should be printed on the console}
}
\value{
a list where the first element is a matrix with random starts, second element all pairwise ARIs
#'
}
\description{
Generate random starts
}
