% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenRatStarts.R
\name{GenRatStarts}
\alias{GenRatStarts}
\title{Title}
\usage{
GenRatStarts(
  DataList,
  nComp,
  nClus,
  scalevalue,
  center,
  verbose,
  pseudo,
  pseudoFac
)
}
\arguments{
\item{DataList}{DataList}

\item{nComp}{number of components}

\item{nClus}{number of clusters}

\item{scalevalue}{value for blockscaling procedure}

\item{center}{center}

\item{verbose}{verbose}

\item{pseudo}{percentage used for perturbation rational starts (between 0)}

\item{pseudoFac}{multiplication factor for pseudo rational starts}
}
\value{
out
}
\description{
Title
}
