% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lir.R
\name{Lir}
\alias{Lir}
\title{Compute loss per data matrix using multivariate regression}
\usage{
Lir(Xi, SkL)
}
\arguments{
\item{Xi}{datablock}

\item{SkL}{list with Group ICA parameters}
}
\description{
Compute loss per data matrix using multivariate regression
}
\keyword{internal}
