% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matcher.R
\name{matcher}
\alias{matcher}
\title{Match components between cluster specific spatial maps}
\usage{
matcher(x, reference, RV = FALSE, ...)
}
\arguments{
\item{x}{object of class CICA}

\item{reference}{integer cluster index that serves as the reference. If nifti path is supplied, clusters will be matched to this template}

\item{RV}{compute modified-RV between cluster components}

\item{...}{other arguments}
}
\value{
out
}
\description{
Match components between cluster specific spatial maps
}
