% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturbation.R
\name{perturbation}
\alias{perturbation}
\title{Perturbate a clustering}
\usage{
perturbation(p, percentage = 0.1)
}
\arguments{
\item{p}{a clustering vector}

\item{percentage}{percentage (in proportions)}
}
\value{
peturbated clustering vector
}
\description{
Perturbate a clustering
}
\keyword{internal}
