% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.CICA.R
\name{plot.CICA}
\alias{plot.CICA}
\title{Plot method for}
\usage{
\method{plot}{CICA}(x, brain = "auto")
}
\arguments{
\item{x}{Object of \code{class} CICA}

\item{brain}{auto}
}
\description{
Plot method for
}
