% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ModSel.R
\name{plot.ModSel}
\alias{plot.ModSel}
\title{Plot method for sequential model selection}
\usage{
\method{plot}{ModSel}(x)
}
\arguments{
\item{x}{Object of \code{class} ModSel}
}
\description{
Plot method for the sequential model selection option for CICA
}
