% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reclus.R
\name{Reclus}
\alias{Reclus}
\title{Title}
\usage{
Reclus(DataList, SrList)
}
\arguments{
\item{DataList}{list with data matrices}

\item{SrList}{list with Group ICA RSN parameters}
}
\value{

}
\description{
Title
}
\keyword{internal}
