% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelSelection.R
\name{SequentialScree}
\alias{SequentialScree}
\title{Sequential Model Selection for Multiple CICA model}
\usage{
SequentialScree(x)
}
\arguments{
\item{x}{an object of class MultipleCICA}
}
\value{
a list object
}
\description{
Sequential Model Selection for Multiple CICA model
}
