% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_CICA.R
\name{Sim_CICA}
\alias{Sim_CICA}
\title{Simulate CICA data}
\usage{
Sim_CICA(Nr, Q, R, voxels, timepoints, E)
}
\arguments{
\item{Nr}{number of subjects per cluster}

\item{Q}{number of components}

\item{R}{number of clusters}

\item{voxels}{number of voxels}

\item{timepoints}{number of time points}

\item{E}{proportion of independent gaussian noise}
}
\value{
a list with simulated CICA data
}
\description{
Simulate CICA data
}
\examples{
Xe <- Sim_CICA(Nr = 15, Q = 5, R = 4, voxels = 100, timepoints = 10, E = .2)
}
