% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sr_to_nifti.R
\name{Sr_to_nifti}
\alias{Sr_to_nifti}
\title{Convert Cluster specific independent components to NIFTI format}
\usage{
Sr_to_nifti(x)
}
\arguments{
\item{x}{an object of \code{class} CICA}
}
\value{
a list with niftiImage files
}
\description{
Convert Cluster specific independent components to NIFTI format
}
