% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xscale.R
\name{xscale}
\alias{xscale}
\title{Scale data blocks to have an equal sum of squares}
\usage{
xscale(Xi, value = 1000)
}
\arguments{
\item{Xi}{a single datablock}

\item{value}{desired sum of squares value of the datablock}
}
\value{

}
\description{
Scale data blocks to have an equal sum of squares
}
\keyword{internal}
