% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CICA.R
\name{CICA}
\alias{CICA}
\title{CICA: Clusterwise Independent Component Analysis}
\usage{
CICA(
  DataList,
  nStarts,
  nComp,
  nClus,
  scale = TRUE,
  center = TRUE,
  rational = NULL,
  maxiter = 100,
  verbose = TRUE
)
}
\arguments{
\item{DataList}{a list of matrices}

\item{nStarts}{number of multiple starts}

\item{nComp}{number of ICA components per cluster}

\item{nClus}{number of clusters}

\item{scale}{scale each matrix to have an equal sum of squares}

\item{center}{mean center matrices}

\item{rational}{optional numeric vector, a (pseudo) rational starting cluster}

\item{maxiter}{maximum number of iterations for each start}

\item{verbose}{print loss information to console}
}
\value{

}
\description{
Main function to perform Clusterwise Independent Component Analysis
}
\author{
Jeffrey Durieux
}
