% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CICA.R, R/summary.CICA.R
\name{CICA}
\alias{CICA}
\alias{summary.CICA}
\title{CICA: Clusterwise Independent Component Analysis}
\usage{
CICA(
  DataList,
  nStarts,
  nComp,
  nClus,
  scale = TRUE,
  center = TRUE,
  rational = NULL,
  maxiter = 100,
  verbose = TRUE
)

\method{summary}{CICA}(object, ...)
}
\arguments{
\item{DataList}{a list of matrices}

\item{nStarts}{number of multiple starts}

\item{nComp}{number of ICA components per cluster}

\item{nClus}{number of clusters}

\item{scale}{scale each matrix to have an equal sum of squares}

\item{center}{mean center matrices}

\item{rational}{optional numeric vector, a (pseudo) rational starting cluster}

\item{maxiter}{maximum number of iterations for each start}

\item{verbose}{print loss information to console}

\item{object}{Object of the type produced by \code{\link{CICA}}}

\item{...}{Additional arguments}
}
\value{
\item{P}{partitioning vector of size \code{length(DataList)}}
\item{Sr}{list of size \code{nClus}, containing cluster specific independent components}
\item{Ais}{list of size \code{length(DataList)}, containing subject specific time courses}
\item{Loss}{loss function value of the best start}
\item{LossStarts}{loss function values of all starts}


}
\description{
Main function to perform Clusterwise Independent Component Analysis
}
\examples{
data('ExampleData', package = 'CICA')
inputdata <- ExampleData$XE
output <- CICA(DataList = inputdata, nStarts = 3, nComp = 5, nClus = 3, verbose = FALSE)
summary(output)
}
\author{
Jeffrey Durieux
}
