% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CICA.R
\name{CICA}
\alias{CICA}
\title{CICA: Clusterwise Independent Component Analysis}
\usage{
CICA(
  DataList,
  nComp,
  nClus,
  RanStarts,
  RatStarts = FALSE,
  pseudo,
  pseudoFac,
  rational = NULL,
  userGrid = NULL,
  scalevalue = NULL,
  center = TRUE,
  maxiter = 100,
  verbose = TRUE,
  ctol = 1e-06
)
}
\arguments{
\item{DataList}{a list of matrices}

\item{nComp}{number or vector of ICA components per cluster}

\item{nClus}{number or vector of clusters}

\item{RanStarts}{number of random starts}

\item{RatStarts}{Generate rational starts}

\item{pseudo}{percentage value for perturbating rational starts to obtain pseudo rational starts}

\item{pseudoFac}{factor to multiply the number of rational starts (7 in total) to obtain pseudorational starts}

\item{rational}{a rational starting seed, if NULL no rational starting seed is used}

\item{userGrid}{user supplied data.frame for multiple model CICA. First column are the requested components. Second column are the requested clusters}

\item{scalevalue}{desired sum of squares of the block scaling procedure}

\item{center}{mean center matrices}

\item{maxiter}{maximum number of iterations for each start}

\item{verbose}{print loss information to console}

\item{ctol}{tolerance value for convergence criterion}
}
\value{
\code{CICA} returns an object of \code{\link{class}} "CICA". It contains the estimated clustering, cluster specific component matrices and subject specific time course matrices
\item{P}{partitioning vector of size \code{length(DataList)}}
\item{Sr}{list of size \code{nClus}, containing cluster specific independent components}
\item{Ais}{list of size \code{length(DataList)}, containing subject specific time courses}
\item{Loss}{loss function value of the best start}
\item{LossStarts}{loss function values of all starts}
}
\description{
Main function to perform Clusterwise Independent Component Analysis
}
\examples{
data('CICA_data', package = 'CICA')
output <- CICA(DataList = CICA_data$X, nStarts = 3, nComp = 5, nClus = 4, verbose = FALSE)
summary(output)
}
\author{
Jeffrey Durieux
}
