% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusf.R
\name{clusf}
\alias{clusf}
\title{Random clustering generation}
\usage{
clusf(nBlocks, nClus)
}
\arguments{
\item{nBlocks}{number of objects}

\item{nClus}{number of clusters}
}
\description{
Random clustering generation
}
\keyword{internal}
