% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeRVmat.R
\name{computeRVmat}
\alias{computeRVmat}
\title{Compute modified RV matrix}
\usage{
computeRVmat(DataList = DataList, dist = TRUE, verbose = TRUE)
}
\arguments{
\item{DataList}{a list with matrices}

\item{dist}{boolean if TRUE distance object is returned}

\item{verbose}{boolean if TRUE progressbar is printed to the console}
}
\value{
\item{RVsS} a matrix or distance object containing the pairwise modified RV values
}
\description{
This function computes a NxN modified RV matrix
}
