% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConcData.R
\name{ConcData}
\alias{ConcData}
\title{Concatenate datablocks into list determined by cluster labels}
\usage{
ConcData(DataList, ClusVec)
}
\arguments{
\item{DataList}{list with datablocks}

\item{ClusVec}{vector with cluster labels}
}
\value{
list with concatenated datablocks
}
\description{
Internal function for CICA package
}
\author{
Jeffrey Durieux
}
\keyword{internal}
