% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractICA.R
\name{ExtractICA}
\alias{ExtractICA}
\title{Extract Group ICA parameters}
\usage{
ExtractICA(DataList, nComp)
}
\arguments{
\item{DataList}{a list object to perform Group ICAs on}

\item{nComp}{number of ICA components to extract}
}
\value{

}
\description{
Extract Group ICA parameters
}
\keyword{internal}
