% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractICA.R
\name{ExtractICA}
\alias{ExtractICA}
\title{Extract Group ICA parameters}
\usage{
ExtractICA(DataList, nComp, svd = TRUE)
}
\arguments{
\item{DataList}{a list object to perform Group ICAs on}

\item{nComp}{number of ICA components to extract}

\item{svd}{}
}
\value{
a list with cluster specific independent components
}
\description{
Internal function for CICA package
}
\keyword{internal}
