% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindRationalStarts.R, R/plot.rstarts.R
\name{FindRationalStarts}
\alias{FindRationalStarts}
\alias{plot.rstarts}
\title{Plot method for rstarts object}
\usage{
FindRationalStarts(
  DataList,
  nComp,
  nClus,
  scale = TRUE,
  center = TRUE,
  verbose = TRUE,
  pseudo = NULL
)

\method{plot}{rstarts}(x, type = 1, mdsdim = 2, ...)
}
\arguments{
\item{DataList}{a list of matrices}

\item{nComp}{number of ICA components to extract}

\item{nClus}{number of clusters}

\item{scale}{scale each matrix to have an equal sum of squares}

\item{center}{mean center matrices}

\item{verbose}{print output to console}

\item{pseudo}{default is \code{NULL}}

\item{x}{an object of \code{class} rstarts}

\item{type}{type of plot, 1 for a dendrogram, 2 for a multidimensional scaling configuration}

\item{...}{optional arguments passed to \code{hclust} function}
}
\value{
dataframe with (pseudo-) rational and dist object based on the pairwise modified RV values
}
\description{
Plot method for rstarts object
}
\examples{
data('CICA_data', package = 'CICA')
rats <- FindRationalStarts(DataList = CICA_data$X, nComp = 5, nClus = 4,verbose = T, pseudo = .2)
plot(rat, type = 1, method = 'ward.D2')
plot(rat, type = 2, method = 'ward.D2')
plot(rat, type = 2, method = 'ward.D2', mdsdim = 3)
}
\references{
Durieux, J., & Wilderjans, T. F. (2019). Partitioning subjects based on high-dimensional fMRI data: comparison of several clustering methods and studying the influence of ICA data reduction in big data. Behaviormetrika, 46(2), 271-311.
}
