% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadNiftis.R
\name{loadNiftis}
\alias{loadNiftis}
\title{Load Nifti files from directory}
\usage{
loadNiftis(dir = ..., toMatrix = TRUE)
}
\arguments{
\item{dir}{Input directory containing nifti files}

\item{toMatrix}{logical if TRUE nifti's are converted to matrices}
}
\value{
list object containing Voxel by Time course matrices
}
\description{
Load Nifti files from directory
}
\examples{
##Do not run
data <- loadNiftis(dir = '/niftifolder/', toMatrix = TRUE)
##End(do not run)
}
