% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reclus.R
\name{Reclus}
\alias{Reclus}
\title{Recluster based on cluster specific components}
\usage{
Reclus(DataList, SrList)
}
\arguments{
\item{DataList}{list with data matrices}

\item{SrList}{list with Group ICA RSN parameters}
}
\value{
\item{newclus}{a list object with an updated partitioning}
\item{SSminVec}{a numeric vector with the lowest loss function values per data block}
\item{Loss}{a numeric indicating the new loss function value}
}
\description{
Internal function for CICA package
}
\keyword{internal}
