% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SearchEmptyClusters.R
\name{SearchEmptyClusters}
\alias{SearchEmptyClusters}
\title{Search for empty clusters}
\usage{
SearchEmptyClusters(nClus, newcluster, SSminVec = NULL)
}
\arguments{
\item{nClus}{number of clusters}

\item{newcluster}{newly updated clustering vector P}

\item{SSminVec}{loss function values per data block}
}
\value{
a numeric vector indicating the new partitioning vector without empty clusters
}
\description{
Internal function for CICA package
}
\keyword{internal}
