% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SearchEmptyClusters.R
\name{SearchEmptyClusters}
\alias{SearchEmptyClusters}
\title{Search for empty clusters}
\usage{
SearchEmptyClusters(nClus, newcluster, SSminVec)
}
\arguments{
\item{newcluster}{newly updated clustering vector P}

\item{oldcluster}{previous clustering vector P}

\item{SSminvec}{loss function values per data block}
}
\value{

}
\description{
Search for empty clusters
}
\keyword{internal}
